Protect against NaN rotation
A simple check against NaN values being set to this.rotation. Without this patch, a NaN rotation value will make a Sprite's texture disappear, which can be confusing. It seems that the rotation should simply not update if a NaN value is passed in, rather than cause other, more confusing side effects.