Fix WebGlContextLost handling.
In WebGlRenderer.js, this.contextLost was being used both as a function
created by bind() and as a boolean variable. Not only is this
incorrect, it also would cause problems in Firefox when calling
removeEventListener, as it would expect contextLost to be a function,
but instead it would get a boolean and would raise an error. Chrome did
not raise an error when calling removeEventListener with a boolean
value as the second argument instead of a function.

This patch separates the two uses of this.contextLost, which is how it
should be.