fix textureUpdate call bug in WebGLRenderer
In PixiShader class, the updateTexture call always uses the
defaultRenderer as reference. However, if the user created more than one
WebGLRenderer instance, it will be wrong.

So, I added an instances array variable under the PIXI.WebGLRenderer and
use the gl.id as index to make sure the updateTexture call works
properly.