RenderTexture - preserve world transform
The renderCanvas and renderWebGL modify the worldTransform, worldAlpha,
and update the transformations of the children. This code adds in a
parameter to the these methods (true by default now, but that can be
changed) to indicate that these transformations should be undone when the
rendering is complete.

While restoring the previous transformations does add a little bit of
overhead this is adventageous in two important ways:

- The default (or requested) behavior works in an "unsurprising manner"
  and does not leave a temporary transformation applied permamently when
  the render operation completes.
- Even if the restore/preserve operation is the documentation and addition
  of method parameter makes this behavior more apparent.