Documentation - Array(Normalization)
YUIDoc does not have a standard (or really any) way of describing inner
types of arrays and there are two general forms used in the current pixi
code:

- `Array<X>`
- `Array An array of X`, or an _inconsistent_ variation of such

The first form is not even parsed correctly by the YUIDoc tool and the
contained type, X, does not appear in the documentation output (or parsed
data) at all. Other similar forms that fail to work are: `X[]`,
`Array{X}`m and `Array.<X>`.

The second form mixes the type (`Array`) and the description of the type
("An array of X"). It would suiting to say "Array of X".

Thus thus commit changes _both_ of the above forms into a normalized
`Array(X)`

While I do not want to impose this as the final chosen representation for
such types, it does _immediately_ address the documentation issues
mentioned above _and_ renders the nested type name in the documentation
output.

If a better method or converntion is found this _consistent_ and
generatable intermim VB-like syntax can be trivially found and replaced.

Another alternative of course, it to forbid anything more complex than
`Array` as a type.