Remove tint from style definition and add as attribute instead.
The Sprite class uses tint as an attribute, and handles any updates that happens to that attribute. For consistency's sake I've removed tint from the style definition, much because the style should pertain to font characteristics. Tint now works in the same way as it does in the Sprite class.