Added better tap detection to the InteractionManager
I’ve added a threshold of number of pixels a touch can move between a
down and up event, before a tap event is emitted.

In the game I’m making right now i have a draggable world, and every
object get’s an tap event when I start&stopping dragging on the object.
Using a threshold prevents is.

Let me know what you think and if it needs adjustment.