Fix for using the loader when loading images
If I use the loader to load images manually and gave them a name, the
images are not added to the texture cache with that name.

We have to use the original url to get them. In could be annoying
because if you load images that could be retina (@2x) you have to acces
them also with the retina url. If I can use the normal url as an
identifier I don’t have to have retina logic to get the right url
spread throughout my code.

For my this fixes this issue, but I don’t know if this will break
existing usage of the loader and cache…