Update package.json main prop
It would be nice if, while developing, it imports the source code, instead the minified one. For example, this code:

```js
import { Container } from 'pixi.js';
```
Using atom and js-hyperclick, you can Ctrl+click `pixi.js` and go into the code. Right now, it opens the minified. With this change, it opens the `src/index.js`.

But the main problem will occur when something goes wrong. Error stacks will be easier to read and trace when using the source code.