Uncomment & fix failing test from #3535.
Resolves #3419 and #3628.

I just saw #3628 after I started working on this and I opened another PR (wrong) about the same bug #3628 looks good to me but adds an extra property, which I personally tend to dislike. This might have some side effects for cases that are untested, but I thought it's worth opening as an alternative to #3628.