Support OffscreenCanvas
The target of InteractionManager could be an OffscreenCanvas, which does
not have features required by the plugin.

Notify the lack of the feature to the plugin initializer and let it ignore
the exception.

createResource checks whether it is an instance of HTMLCanvasElement to see
if it is a canvas. Check also whether it is an instance of OffscreenCanvas.

Now OffscreenCanvas is supported throughout PixiJS. Update documents
accordingly.