Removes default loadType from the spritesheetParser load options (#4573)
Assuming a loadType of Resource.LOAD_TYPE.IMAGE conflicts with the ability to load an image which can't be loaded via the <image> tag.  This happens in the case of loading a compressed image which should be handled by the XmlHttpRequest file loader as an array buffer type.  In the resource loader, there is already a lookup to determine the load type: https://github.com/englercj/resource-loader/blob/master/src/Resource.js#L126
In the case of a binary image asset, the spritesheetParser's loadType stomps on a dynamic lookup based on filetype.