Bugfix: Don't autoclose paths when using canvas (#4154)
When defining a path, any call to beginFill results in the path being closed,
but only for the Canvas renderer, not for the WebGL renderer.
This commit removes the check that sets the shape to be closed when a fill is
active. This unifies the rendering behavior between Canvas and WebGL.