Fixes depthTexture to work
This is the start of  enabling depthTexture support.

This fixes the depthTexture from not being written to at all, and ads a depthResource that is used to fix the internalType needed for webgl2.0

This is a first pass at it, to get it working fully in webgl2.0. Then Ill move to 1.0 + extensions.

The depth resource is also something I'd like to instead be a option we can pass into either Resource itself as a property (internalFormat) or as a property for baseTexture. For now tho it'll have to do.