Use only sourceFrame to calculate projectionMatrix
The projection-system should produce a matrix that converts
world-space coordinates into clip-space coordinates.

The sourceFrame is the rectangle in the scene-graph
(or world) we want to render. It can also be thought of the
"screen" in world-space coordinates.

The clip-space rectangle is always x = -1, 1; y = -1, 1.

The destination frame is just another word for the viewport -
the screen rectangle inside the canvas.