Bind to null framebuffer only when copy from main canvas (#10486)
The render target system's copyToTexture assumes that if the source render surface is a canvas, then it must be the main framebuffer. But this is false if someone is trying to copy from a secondary or offscreen canvas.

For a bug repro, see this JSFiddle
https://jsfiddle.net/ShukantPal/vwrh51qs/27/

I've verified locally that this fix causes the gradient to paint correctly