Fix: Correct color format in `addColorStop` function for Android WebView compatibility
Android WebView does not support the `#RRGGBBAA` color format in the `addColorStop` method. Using this format results in the following error:

```
SyntaxError: Failed to execute 'addColorStop' on 'CanvasGradient': The value provided ('#000000ff') could not be parsed as a color.
```
For further details, refer to the Chromium issue tracker: [Chromium Issue #40222657](https://issues.chromium.org/issues/40222657).