fix: DOMContainer scaling if canvas width/height is percentage (#11539)
* fix: DOMContainer scaling if canvas width/height is percentage

Addresses issues with incorrect DOMContainer positioning
when the canvas is scaled.

Introduces CanvasTransformSync to manage the DOM element's
transform based on canvas size and position, utilizing
ResizeObserver for efficient updates. Uses a fallback to the Ticker is ResizeObserver is not found (been supported since 2020)

Ensures correct scaling and translation of DOM elements
relative to the canvas, resolving discrepancies in rendering.

* PR feedback

* fix offscreen canvas