fix: Installed bitmap text showing wrong fill (#11523)
* fix: Installed bitmap text showing wrong fill

Fixes: #11440

Updates the logic used to determine when bitmap text styles can use tinting instead of baking colors.

Includes a new visual test case to verify correct rendering of bitmap text with gradient fills and tinting.

* add allowFillOverride

* Renames `allowFillOverride` to `dynamicFill`

Renames the `allowFillOverride` option to `dynamicFill` for clarity and consistency.

This change provides a more descriptive name that better reflects the option's purpose, which is to control whether the fill color of a bitmap font can be dynamically changed via tinting.